+++ packages/ai/CHANGELOG.md
@@ -1,5 +1,17 @@
 # ai
 
+## 4.1.52
+
+### Patch Changes
+
+- 5329a69: fix (ai/core): fix duplicated reasoning in streamText onFinish and messages
+
+## 4.1.51
+
+### Patch Changes
+
+- 0cb2647: feat (ai/core): add streamText sendStart & sendFinish data stream options
+
 ## 4.1.50
 
 ### Patch Changes