+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,20 @@
 
 ### Patch Changes
 
+- Updated dependencies [5329a69]
+  - ai@4.1.52
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [0cb2647]
+  - ai@4.1.51
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [ae98f0d]
   - ai@4.1.50