+++ packages/valibot/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/valibot
 
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies [5329a69]
+  - ai@4.1.52
+
+## 0.0.18
+
+### Patch Changes
+
+- Updated dependencies [0cb2647]
+  - ai@4.1.51
+
 ## 0.0.17
 
 ### Patch Changes